Remove unsafe Promise operations
Functions Promise.prototype.{each, map, filter, reduce, spread} (not to
be mistaken with "static" Promise.each, Promise.map etc) are not
type-safe in a sense that they are only defined for Promisees which
contain arrays but this behavior cannot be expressed for type-checked
so the result type of them is effectively `any`.

We remove them as they introduce dangerous gaps in type safety and also
as a part of departing from Bluebird.